Major Redo of Sonarr V3 Release Profile
Major Redo of Sonarr V3 Release Profile
I decided to redo the Release Profile to make it easier for for me and the end user to update/manage the Release Profile, I decided not to add the scene groups anymore (and also removing the current ones). Why ? Lately I noticed allot of popup scene releases meaning we need to update often.**
  **With this it also means all scene and (yet) unknown groups will be getting a score of `0` and not `-50`.
I will still keep a eye on them and add the Low Quality Groups to the `-100` score and when I find a new P2P group I will add them, resulting in less updates and a more accurate upgrade.