Reviewed PR and transferred changes to Radarr
```yml
Updated: Sonarr Recommended naming scheme
- Changes: Clarify grammar PR
```
```yml
Updated: Radarr Recommended naming scheme
- Changes: Clarify grammar PR
```
```yml
Security Fixes: prevent a type of phishing known as tabnabbing.
- Added: noopener noreferrer to external URL redirects (others pages will follow soon)
  - Info: Tabnabbing, sometimes called reverse tabnabbing, is an exploit that uses the browser's default behavior with target="_blank" to gain partial access to your page through the window.object API.

With tabnabbing, a page that you link to could cause your page to redirect to a fake login page. This would be hard for most users to notice because the focus would be on the tab that just opened – not the original tab with your page.
```