NinjaCentral fix (SABnzbd)
This is a fix related to how NinjaCentral names their NZB files.
All their NZBs have a dash + four characters appended to the filename/name of the NZB.
A few examples: 
-ASn1 
-xm98 
-ga2k

This reg-ex/reg-ex-replace works by removing that part only. I've had this reg-ex in my clean.py script for quite a long time, and haven't observed any negative effects with using this, with a lot of different NZB indexers simultaneously. 

A reg-ex test:
https://regex101.com/r/fpsQqh/1