Update radarr-naming.json
This seems like it fixes an issue where AudioCodec / AudioChannels might come back empty. If this is a desired change it'd be for all the other profiles below it. Looking for review on this though.

Current suggestion:
```
{Movie CleanTitle} {(Release Year)} {imdb-{ImdbId}} {edition-{Edition Tags}} {[Custom Formats]}{[Quality Full]}{[MediaInfo 3D]}{[MediaInfo VideoDynamicRangeType]}{[Mediainfo AudioCodec}{ Mediainfo AudioChannels}][{Mediainfo VideoCodec}]{-Release Group}
Foobar (1999) {imdb-tt12345678} [WEBRip-720p]][]-Group.mp4
```

My suggestion:
```
{Movie CleanTitle} {(Release Year)} {imdb-{ImdbId}} {edition-{Edition Tags}} {[Custom Formats]}{[Quality Full]}{[MediaInfo 3D]}{[MediaInfo VideoDynamicRangeType]}[{Mediainfo AudioCodec}{ Mediainfo AudioChannels}][{Mediainfo VideoCodec}]{-Release Group}
Foobar (1999) {imdb-tt12345678} [WEBRip-720p][][]-Group.mp4
```

Sadly I cannot figure out a way to NOT include the empty portions. Seems like that's up to the *arr to fix how it does string/token replacements.